+++ views/consultation/rest_get.php
@@ -3,22 +3,40 @@
 /** @var \app\models\db\Consultation $consultation */
 
 use app\components\UrlHelper;
-use app\models\db\{Amendment, Motion};
+use app\models\db\{Amendment, IMotion, Motion};
 
 $json = [
     'title' => $consultation->title,
     'title_short' => $consultation->titleShort,
-    'motion_links' => array_map(function (Motion $motion) {
+    'motion_links' => array_map(function (IMotion $imotion) {
+        if (is_a($imotion, Amendment::class)) {
+            $title = $imotion->getTitle();
+            $titleWithIntro = $imotion->getTitle();
+            $amendments = [];
+            $type = 'amendment';
+            $htmlLink = UrlHelper::absolutizeLink(UrlHelper::createAmendmentUrl($imotion));
+            $jsonLink = UrlHelper::absolutizeLink(UrlHelper::createAmendmentUrl($imotion, 'rest'));
+        } else {
+            /** @var Motion $imotion */
+            $title = $imotion->title;
+            $titleWithIntro = $imotion->getTitleWithPrefix();
+            $amendments = $imotion->getVisibleAmendmentsSorted();
+            $type = 'motion';
+            $htmlLink = UrlHelper::absolutizeLink(UrlHelper::createMotionUrl($imotion));
+            $jsonLink = UrlHelper::absolutizeLink(UrlHelper::createMotionUrl($imotion, 'rest'));
+        }
+        /** @var IMotion $imotion */
         return [
-            'id' => $motion->id,
-            'agenda_item' => ($motion->agendaItem ? $motion->agendaItem->title : null),
-            'prefix' => $motion->titlePrefix,
-            'title' => $motion->title,
-            'title_with_intro' => $motion->getTitleWithIntro(),
-            'title_with_prefix' => $motion->getTitleWithPrefix(),
-            'status_id' => $motion->status,
-            'status_title' => $motion->getFormattedStatus(),
-            'initiators_html' => $motion->getInitiatorsStr(),
+            'type' => $type,
+            'id' => $imotion->id,
+            'agenda_item' => ($imotion->agendaItem ? $imotion->agendaItem->title : null),
+            'prefix' => $imotion->titlePrefix,
+            'title' => $title,
+            'title_with_intro' => $titleWithIntro,
+            'title_with_prefix' => $imotion->getTitleWithPrefix(),
+            'status_id' => $imotion->status,
+            'status_title' => $imotion->getFormattedStatus(),
+            'initiators_html' => $imotion->getInitiatorsStr(),
             'amendment_links' => array_map(function (Amendment $amendment) {
                 return [
                     'id' => $amendment->id,
@@ -29,9 +47,9 @@ $json = [
                     'url_json' => UrlHelper::absolutizeLink(UrlHelper::createAmendmentUrl($amendment, 'rest')),
                     'url_html' => UrlHelper::absolutizeLink(UrlHelper::createAmendmentUrl($amendment)),
                 ];
-            }, $motion->getVisibleAmendmentsSorted()),
-            'url_json' => UrlHelper::absolutizeLink(UrlHelper::createMotionUrl($motion, 'rest')),
-            'url_html' => UrlHelper::absolutizeLink(UrlHelper::createMotionUrl($motion)),
+            }, $amendments),
+            'url_json' => $jsonLink,
+            'url_html' => $htmlLink,
         ];
     }, $consultation->getVisibleIMotionsSorted(false)),
     'url_json' => UrlHelper::absolutizeLink(UrlHelper::createUrl('consultation/rest')),