+++ config/urls.php
@@ -88,6 +88,7 @@ $urlRules = [
 
     $restBaseCon . '/votings/open' => '/voting/get-open-voting-blocks',
     $restBaseCon . '/votings/<votingBlockId:[^\/]+>/<itemType:[^\/]+>/<itemId:[^\/]+>' => '/voting/post-vote',
+    $restBaseCon . '/votings/<votingBlockId:[^\/]+>/<itemType:[^\/]+>/<itemId:[^\/]+>/settings' => '/voting/post-vote-settings',
 
     $domv . 'motion/pdfcollection/<motionTypeId:\d+>/<filename:.*>' => 'motion/pdfcollection',
     $domv . 'motion/fullpdf/<motionTypeId:\d+>/<filename:.*>'       => 'motion/fullpdf',