+++ messages/de/voting.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'bc' => 'Abstimmungen',
+
+    'admin_title' => 'Abstimmungen: Verwaltung',
+    'admin_bc' => 'Verwaltung',
+    'admin_voting_use' => 'Online-Abstimmung über Antragsgrün durchführen',
+];