+++ messages/en/voting.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'bc' => 'Votings',
+
+    'admin_title' => 'Voting administration',
+    'admin_bc' => 'Administration',
+    'admin_voting_use' => 'Use Antragsgrün for online voting',
+];