+++ views/voting/admin-votings.php
@@ -1,6 +1,6 @@
 <?php
 
-use app\models\db\VotingBlock;
+use app\components\UrlHelper;
 use app\models\proposedProcedure\Factory;
 use yii\helpers\Html;
 
@@ -13,7 +13,9 @@ $controller = $this->context;
 /** @var \app\models\db\Consultation */
 $consultation = $controller->consultation;
 $layout       = $controller->layoutParams;
-$layout->addBreadcrumb('Votings');
+$layout->addBreadcrumb(Yii::t('voting', 'bc'));
+$layout->addBreadcrumb(Yii::t('voting', 'admin_bc'));
+$this->title = Yii::t('voting', 'admin_title');
 
 $layout->loadVue();
 $layout->addVueTemplate('@app/views/voting/admin-votings.vue.php');
@@ -21,25 +23,28 @@ $layout->addVueTemplate('@app/views/voting/admin-votings.vue.php');
 $proposalFactory = new Factory($consultation, false);
 $apiData = [];
 foreach ($proposalFactory->getAllVotingBlocks() as $votingBlock) {
+    /** @noinspection PhpUnhandledExceptionInspection */
     $apiData[] = $votingBlock->getAdminApiObject();
 }
 
+$voteSettingsUrl = UrlHelper::createUrl(['/voting/post-vote-settings', 'votingBlockId' => 'VOTINGBLOCKID', 'itemType' => 'ITEMTYPE', 'itemId' => 'ITEMID']);
+
 ?>
-<h1>Voting administration</h1>
+<h1><?= Yii::t('voting', 'admin_title') ?></h1>
 <div class="content">
     ...
 </div>
-<section >
-
-                </section>
 <div class="manageVotings votingCommon"
+     data-url-vote-settings="<?= Html::encode($voteSettingsUrl) ?>"
      data-antragsgruen-widget="backend/VotingAdmin"
      data-voting="<?= Html::encode(json_encode($apiData)) ?>">
     <div class="votingAdmin"></div>
 </div>
+<!--
 <div class="content votingAdderForm">
     <button class="btn btn-link btnAddOpener" type="button">
         <span class="glyphicon glyphicon-plus-sign" aria-hidden="true"></span>
         Create a new voting
     </button>
 </div>
+-->